Rename app namespace
