+++ nearby-app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId "com.esri.android.nearbyplaces"
+        applicationId "com.esri.arcgisruntime.opensourceapps.nearbyplaces"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
