+++ nearby-app/src/main/java/com/esri/arcgisruntime/opensourceapps/nearbyplaces/data/Place.java
@@ -21,7 +21,7 @@
  * email: contracts@esri.com
  *
  */
-package com.esri.android.nearbyplaces.data;
+package com.esri.arcgisruntime.opensourceapps.nearbyplaces.data;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
