+++ nearby-app/src/main/java/com/esri/arcgisruntime/opensourceapps/nearbyplaces/places/PlacesContract.java
@@ -22,12 +22,12 @@
  *
  */
 
-package com.esri.android.nearbyplaces.places;
+package com.esri.arcgisruntime.opensourceapps.nearbyplaces.places;
 
 import android.location.Location;
-import com.esri.android.nearbyplaces.BasePresenter;
-import com.esri.android.nearbyplaces.BaseView;
-import com.esri.android.nearbyplaces.data.Place;
+import com.esri.arcgisruntime.opensourceapps.nearbyplaces.BasePresenter;
+import com.esri.arcgisruntime.opensourceapps.nearbyplaces.BaseView;
+import com.esri.arcgisruntime.opensourceapps.nearbyplaces.data.Place;
 import com.esri.arcgisruntime.geometry.Envelope;
 
 import java.util.List;
