+++ nearby-app/src/main/res/menu/map_menu.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.esri.android.nearbyplaces.MainActivity">
+      tools:context="com.esri.opensourceapps.nearbyplaces.MainActivity">
     <item android:id="@+id/list_action"
           android:title="@string/list_view"
           android:icon="@drawable/ic_list_white_24px"
