+++ nearby-app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.esri.android.nearbyplaces.MainActivity">
+      tools:context="com.esri.opensourceapps.nearbyplaces.MainActivity">
     <item android:id="@+id/map_action"
           android:title="@string/map_view"
           android:icon="@drawable/ic_map_white_24px"
